Fix deletion of symlink to temporary directory
Bug fix: an additional check confirming that 'var.tlink' is a symlink should be carried out before deletion in order to avoid ungraceful termination, as the symbolic link ('var.tlink') to the temporary directory ('var.temp') may not created when the current working directory ('var.home') is the same as the temporary directory's parent ('/tmp').

+++ pkg/libs/Tools.py
@@ -115,7 +115,7 @@ class Tools:
         os.chdir(var.home)
 
         # Removes the temporary link created at the start of the app
-        if os.path.exists(var.tlink):
+        if os.path.islink(var.tlink) and os.path.exists(var.tlink):
             os.remove(var.tlink)
 
             if os.path.exists(var.tlink):
